requirements.txt
# pip freeze > requirements.txt
# pip install -r requirements.txt

psycopg==3.0.1
psycopg-binary==3.0.1
python-dotenv==0.19.1
